Adds a comparator to sort events by
This allows the events to be displayed in chronological order. The list is continuously sorted every time a new event is added. The consideration to use a binary search tree was made, although I believe that it's unneccesary and we can simply stick with our usual representation for consistency sake and also for the fact that the number of events will be small. This idea may be revisited in the future.